+++ beets/ui/commands.py
@@ -23,6 +23,7 @@ import time
 import itertools
 import codecs
 import platform
+import re
 
 import beets
 from beets import ui
@@ -110,6 +111,30 @@ fields_cmd = ui.Subcommand(
 )
 fields_cmd.func = fields_func
 default_commands.append(fields_cmd)
+
+
+# help: Print help text for commands
+
+class HelpCommand(ui.Subcommand):
+
+    def __init__(self):
+        super(HelpCommand, self).__init__(
+            'help', aliases=('?',),
+            help='give detailed help on a specific sub-command',
+        )
+
+    def func(self, lib, opts, args):
+        if args:
+            cmdname = args[0]
+            helpcommand = self.root_parser._subcommand_for_name(cmdname)
+            if not helpcommand:
+                raise ui.UserError("unknown command '{0}'".format(cmdname))
+            helpcommand.print_help()
+        else:
+            self.root_parser.print_help()
+
+
+default_commands.append(HelpCommand())
 
 
 # import: Autotagger and importer.
@@ -1498,7 +1523,8 @@ def completion_script(commands):
         command_names.append(name)
 
         for alias in cmd.aliases:
-            aliases[alias] = name
+            if re.match(r'^\w+$', alias):
+                aliases[alias] = name
 
         options[name] = {'flags': [], 'opts': []}
         for opts in cmd.parser._get_all_options()[1:]:
@@ -1516,9 +1542,6 @@ def completion_script(commands):
         'flags': ['-v', '--verbose'],
         'opts': '-l --library -c --config -d --directory -h --help'.split(' ')
     }
-
-    # Help subcommand
-    command_names.append('help')
 
     # Add flags common to all commands
     options['_common'] = {